pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Miyamura80", email = "eitomiyamura@gmail.com" }
]
dependencies = [
    "black>=24.8.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "human-id>=0.2.0",
    "pytest>=8.3.3",
    "google-api-python-client>=2.146.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/python_template"]
